Changed the PV type for the checksum PVs in cps_recv to double to match daqd.
Daqd exports all PVs as doubles, changing the channel list checksum and data checksum to be doubles to match.  If this is not done, then a negative value is shown if the high bit is set when viewing the cps_recv version of the checksum, where the daqd shows a large positive number.

+++ src/pub_sub_stream/dc_stats.hh
@@ -147,8 +147,8 @@ private:
     int          not_stalled_{ 0 };
     int          uptime_{ 0 };
     unsigned int gpstime_{ 0 };
-    int          data_crc_{ 0 };
-    int          channel_config_hash_{ 0 };
+    double       data_crc_{ 0 };
+    double       channel_config_hash_{ 0 };
     int          unique_dcus_per_sec_{ 0 };
     int          data_rate_{ 0 };
     int          total_chans_{ 0 };
